+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Unreleased: changes on master, not yet released
 ### Changed
 
  - [#2923][] Updated member variables of `RigidBodyLoop` and `RigidBodActuator` to conform to style guide.
+ - [#2905][] Made `RigidBody::body_index` private. Added necessary accessors.
  - [#2904][] Made `RigidBody::parent` private. Re-named it to be `RigidBody::parent_`. Added necessary accessors.
  - [#2903][] Made `RigidBody::robotnum` private. Re-named it to be `RigidBody::model_id_`.
  - [#2902][] Made `RigidBody::model_name_` private. Re-named `RigidBody::model_name()` to be `RigidBody::get_model_name()`. Added `RigidBody::set_model_name()`.
@@ -92,5 +93,6 @@ Changes in version v0.9.11 and before are not provided.
 [#2902]: https://github.com/RobotLocomotion/drake/issues/2902
 [#2903]: https://github.com/RobotLocomotion/drake/issues/2903
 [#2904]: https://github.com/RobotLocomotion/drake/issues/2904
+[#2905]: https://github.com/RobotLocomotion/drake/issues/2905
 [#2923]: https://github.com/RobotLocomotion/drake/issues/2923
 [#2963]: https://github.com/RobotLocomotion/drake/issues/2963